started renaming herring to item
git-svn-id: svn+ssh://svn.code.sf.net/p/supertuxkart/code/trunk/supertuxkart@2389 178a84e3-b1eb-0310-8ba1-8eac791a3b58

+++ src/highscore_manager.cpp
@@ -102,7 +102,7 @@ void HighscoreManager::Load()
         
         // check file version
         int v;
-        if (!node->get("file-version",v) || v<CURRENT_HSCORE_FILE_VERSION)
+        if (!node->get("file-version",v) || v<(int)CURRENT_HSCORE_FILE_VERSION)
         {
             fprintf(stderr, "Highscore file format too old, a new one will be created.\n");
             std::string warning = _("The highscore file was too old,\nall highscores have been erased.");
